# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Small Town Watchdog API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sources:
    get:
      operationId: listSources
      tags: [sources]
      summary: List source URLs
      parameters:
        - name: entityId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Source"
    post:
      operationId: createSource
      tags: [sources]
      summary: Create a source URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
  /sources/{id}:
    patch:
      operationId: updateSource
      tags: [sources]
      summary: Update a source URL
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourcePatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "404":
          description: Not found
    delete:
      operationId: deleteSource
      tags: [sources]
      summary: Delete a source URL
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    Source:
      type: object
      required: [id, entityId, url, name, sourceType, isActive, createdAt]
      properties:
        id:
          type: integer
        entityId:
          type: integer
        url:
          type: string
        name:
          type: string
        sourceType:
          type: string
        isActive:
          type: boolean
        lastChecked:
          type: ["string", "null"]
        createdAt:
          type: string
    SourceInput:
      type: object
      required: [entityId, url, name, sourceType]
      properties:
        entityId:
          type: integer
        url:
          type: string
        name:
          type: string
        sourceType:
          type: string
        isActive:
          type: boolean
    SourcePatch:
      type: object
      properties:
        url:
          type: string
        name:
          type: string
        sourceType:
          type: string
        isActive:
          type: boolean
...